model/LLM/requirements.txt
python-chess
pandas
datasets
transformers
peft
trl
bitsandbytes
